+++ src/Commands/Admin/SetTenant.cs
@@ -274,9 +274,6 @@ namespace PnP.PowerShell.Commands.Admin
         public bool? EnableAzureADB2BIntegration;
 
         [Parameter(Mandatory = false)]
-        public bool? SyncAadB2BManagementPolicy;
-
-        [Parameter(Mandatory = false)]
         public bool? CoreRequestFilesLinkEnabled;
 
         [Parameter(Mandatory = false)]
@@ -287,9 +284,6 @@ namespace PnP.PowerShell.Commands.Admin
 
         [Parameter(Mandatory = false)]
         public bool? DisableDocumentLibraryDefaultLabeling { get; set; }
-
-        [Parameter(Mandatory = false)]
-        public bool? DisableListSync { get; set; }
 
         [Parameter(Mandatory = false)]
         public bool? IsEnableAppAuthPopUpEnabled { get; set; }
@@ -382,6 +376,31 @@ namespace PnP.PowerShell.Commands.Admin
 
         [Parameter(Mandatory = false)]
         public bool? SiteOwnerManageLegacyServicePrincipalEnabled { get; set; }
+
+        [Parameter(Mandatory = false)]
+        public bool? ReduceTempTokenLifetimeEnabled { get; set; }
+        
+        [Parameter(Mandatory = false)]
+        public int? ReduceTempTokenLifetimeValue;
+
+        [Parameter(Mandatory = false)]
+        public bool? ViewersCanCommentOnMediaDisabled { get; set; }
+        
+        [Parameter(Mandatory = false)]
+        public bool? AllowGuestUserShareToUsersNotInSiteCollection { get; set; }
+
+        [Parameter(Mandatory = false)]
+        public string ConditionalAccessPolicyErrorHelpLink;
+
+        [Parameter(Mandatory = false)]
+        public string CustomizedExternalSharingServiceUrl;
+
+        [Parameter(Mandatory = false)]
+        public bool? IncludeAtAGlanceInShareEmails { get; set; }
+
+        [Parameter(Mandatory = false)]
+        public bool? MassDeleteNotificationDisabled { get; set; }
+
         protected override void ExecuteCmdlet()
         {
             AdminContext.Load(Tenant);
@@ -403,6 +422,13 @@ namespace PnP.PowerShell.Commands.Admin
                 Tenant.NoAccessRedirectUrl = NoAccessRedirectUrl;
                 modified = true;
             }
+
+            if (CustomizedExternalSharingServiceUrl != null)
+            {
+                Tenant.CustomizedExternalSharingServiceUrl = CustomizedExternalSharingServiceUrl;
+                modified = true;
+            }
+
             if (ExternalServicesEnabled.HasValue)
             {
                 Tenant.ExternalServicesEnabled = ExternalServicesEnabled.Value;
@@ -1082,7 +1108,7 @@ namespace PnP.PowerShell.Commands.Admin
             {
                 if (OneDriveRequestFilesLinkExpirationInDays.Value < 0 || OneDriveRequestFilesLinkExpirationInDays.Value > 730)
                 {
-                    throw new PSArgumentException($"{OneDriveRequestFilesLinkExpirationInDays} must have a value between 0 and 730", nameof(OneDriveRequestFilesLinkExpirationInDays));
+                    throw new PSArgumentException("OneDriveRequestFilesLinkExpirationInDays must have a value between 0 and 730", nameof(OneDriveRequestFilesLinkExpirationInDays));
                 }
 
                 Tenant.OneDriveRequestFilesLinkExpirationInDays = OneDriveRequestFilesLinkExpirationInDays.Value;
@@ -1092,12 +1118,6 @@ namespace PnP.PowerShell.Commands.Admin
             if (EnableRestrictedAccessControl.HasValue)
             {
                 Tenant.EnableRestrictedAccessControl = EnableRestrictedAccessControl.Value;
-                modified = true;
-            }
-
-            if (SyncAadB2BManagementPolicy.HasValue)
-            {
-                Tenant.SyncAadB2BManagementPolicy = SyncAadB2BManagementPolicy.Value;
                 modified = true;
             }
 
@@ -1117,10 +1137,21 @@ namespace PnP.PowerShell.Commands.Admin
             {
                 if (CoreRequestFilesLinkExpirationInDays.Value < 0 || CoreRequestFilesLinkExpirationInDays > 730)
                 {
-                    throw new PSArgumentException($"{CoreRequestFilesLinkExpirationInDays} must have a value between 0 and 730", nameof(CoreRequestFilesLinkExpirationInDays));
+                    throw new PSArgumentException("CoreRequestFilesLinkExpirationInDays must have a value between 0 and 730", nameof(CoreRequestFilesLinkExpirationInDays));
                 }
 
                 Tenant.CoreRequestFilesLinkExpirationInDays = CoreRequestFilesLinkExpirationInDays.Value;
+                modified = true;
+            }
+
+            if (ReduceTempTokenLifetimeValue.HasValue)
+            {
+                if (ReduceTempTokenLifetimeValue.Value < 5 || ReduceTempTokenLifetimeValue > 15)
+                {
+                    throw new PSArgumentException("ReduceTempTokenLifetimeValue must have a value between 5 and 15", nameof(ReduceTempTokenLifetimeValue));
+                }
+
+                Tenant.ReduceTempTokenLifetimeValue = ReduceTempTokenLifetimeValue.Value;
                 modified = true;
             }
 
@@ -1130,15 +1161,15 @@ namespace PnP.PowerShell.Commands.Admin
                 modified = true;
             }
 
-            if (DisableDocumentLibraryDefaultLabeling.HasValue)
+            if (ConditionalAccessPolicyErrorHelpLink != null)
             {
-                Tenant.DisableDocumentLibraryDefaultLabeling = DisableDocumentLibraryDefaultLabeling.Value;
+                Tenant.ConditionalAccessPolicyErrorHelpLink = ConditionalAccessPolicyErrorHelpLink;
                 modified = true;
             }
 
-            if (DisableListSync.HasValue)
+            if (DisableDocumentLibraryDefaultLabeling.HasValue)
             {
-                Tenant.DisableListSync = DisableListSync.Value;
+                Tenant.DisableDocumentLibraryDefaultLabeling = DisableDocumentLibraryDefaultLabeling.Value;
                 modified = true;
             }
 
@@ -1271,6 +1302,36 @@ namespace PnP.PowerShell.Commands.Admin
             if (SiteOwnerManageLegacyServicePrincipalEnabled.HasValue)
             {
                 Tenant.SiteOwnerManageLegacyServicePrincipalEnabled = SiteOwnerManageLegacyServicePrincipalEnabled.Value;
+                modified = true;
+            }
+
+            if (ReduceTempTokenLifetimeEnabled.HasValue)
+            {
+                Tenant.ReduceTempTokenLifetimeEnabled = ReduceTempTokenLifetimeEnabled.Value;
+                modified = true;
+            }
+
+            if (ViewersCanCommentOnMediaDisabled.HasValue)
+            {
+                Tenant.ViewersCanCommentOnMediaDisabled = ViewersCanCommentOnMediaDisabled.Value;
+                modified = true;
+            }
+
+            if (AllowGuestUserShareToUsersNotInSiteCollection.HasValue)
+            {
+                Tenant.AllowGuestUserShareToUsersNotInSiteCollection = AllowGuestUserShareToUsersNotInSiteCollection.Value;
+                modified = true;
+            }
+            
+            if (IncludeAtAGlanceInShareEmails.HasValue)
+            {
+                Tenant.IncludeAtAGlanceInShareEmails = IncludeAtAGlanceInShareEmails.Value;
+                modified = true;
+            }
+            
+            if (MassDeleteNotificationDisabled.HasValue)
+            {
+                Tenant.MassDeleteNotificationDisabled = MassDeleteNotificationDisabled.Value;
                 modified = true;
             }