+++ src/Commands/Files/GetFile.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace PnP.PowerShell.Commands.Files
 {
-    [Cmdlet(VerbsCommon.Get, "PnPFile", DefaultParameterSetName = "Return as file object")]
+    [Cmdlet(VerbsCommon.Get, "PnPFile", DefaultParameterSetName = URLASFILEOBJECT)]
     public class GetFile : PnPWebCmdlet
     {
         private const string URLTOPATH = "Save to local path";